[raw data] Only include current columns in latest view query (Recidiviz/recidiviz-data#37027)
## Description of the change

modify latest view query so we are not pulling deleted columns into the
view. i ultimately want to get to a state where when a column is marked
as deleted we are free to delete it from BQ, but because we currently
have so many columns that are marked as deleted that have not been
deleted from BQ, it's hard to align our update bq schema logic.

i had to add a description to
`recidiviz/ingest/direct/regions/us_nd/raw_data/us_nd_docstars_ftr_plan_objective.yaml`
because the only column with a description was marked as deleted, so the
view generation was failing. the other markdown changes are just a
result of regenerating the documentation.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#35844

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 81f942cd9af0ecac7f45fe4bd090c16ecb290108