[kpis] add hard failure spans (Recidiviz/recidiviz-data#37058)
## Description of the change

adds a view to `platform_kpis` dataset that will underpin the two
following KPIs
- `Validation hard failure time to resolution in prod`
- `Validation hard failures / month in prod`

this PR adds the `validation_hard_failure_spans` view, which computes
the start and end windows of each hard failure for any validation; the
status end window of a hard failure (`validation_resolution_status`) can
be :
- `SUCCESS`  -- we started "succeeding"
- `FAIL_SOFT` -- we moved to "fail soft"
- `REMOVED` -- the last seen status was `FAIL_HARD`, but we are no
longer running this validation. the `resolution_datetime` will be the
last `run_datetime`
- `ONGOING` -- the last seen status was `FAIL_HARD` and it was seen
during the most recent validation run. the `resolution_datetime` will be
`NULL`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#36971

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 33c538f53c5162878b61a879b3fad9f2cacb2944