[raw data] two small updates to `compare_raw_data` (Recidiviz/recidiviz-data#37503)
## Description of the change

this script has been SO SO helpful, tysm @emilyemilyemilyemilyemilyemily
for writing it! love the design & decomp of the various parts, thank you
for building amazing tools

this PR makes two small updates to it that i found helpful while
validating tennessee rollout, both related to the `--save-to-table`
flag:
1. if i run the script twice, it would error bc of overwriting an
existing table. instead of adding the `overwrite=True` flag to the
`create_table_from_query_async`, i figured it would be nice to have each
run be unique so you could see how things changed between two runs;
thus, i added the unix timestamp of the run to the table prefix so each
run will write to a new table.
2. the script will create output tables for all file tags, even if there
wasn't a difference -- and with the 500 row preview it printed out past
where i could scroll up in my terminal; to figure out which tables had
differences, i needed to click through all the blank tables to figure
out which ones had data. to remedy this, i added a delete clause to
delete the table if the run was successful, so there will only be tables
where there were errors (i.e. that have data).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: bbe76df5f9ba598db05235c0c0c84b861781af36