[Source Tables] Some small SourceTableCollectionUpdateConfig terminology updates (Recidiviz/recidiviz-data#37578)
## Description of the change

Updates the naming on a few of the `SourceTableCollectionUpdateConfig`
constructors for clarity:
* `static()` -> `protected()` - we use this config for raw data tables
and a number of other tables that aren't static at all, but rather are
managed but not easily regenerable and need to be protected from changes
that might produce data loss.
* `unmanaged()` -> `externally_managed()` - I made this name update to
match the terminology we use elsewhere for these tables.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#37577

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0850e68cf93d30fffdb359d8d0925740cbebb5d3