Allow GCS exports to projects outside of data platform projects (Recidiviz/recidiviz-data#37597)
## Description of the change

This PR adds a field to `ExportViewCollectionConfig` to the map data
platform projects, i.e. `recidiviz-123` and `recidiviz-staging`, to the
corresponding projects outside of the data platform, i.e.
`recidiviz-dashboards-production` and `recidiviz-dashboards-staging`
respectively. See https://github.com/Recidiviz/recidiviz-data/pull/37598
for usage IRL

In Recidiviz/recidiviz-data#37260, we want to export product views to a GCS bucket managed in
the Dashboards projects (code for instantiating buckets with Terraform
[here](https://github.com/Recidiviz/recidiviz-dashboards/pull/7156)). So
I needed a way to override using the project id of the Compute Engine
with the desired output project

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#37260

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d8a72846c403cf520c3f550cb321a8ff075f62f0